refactor: simplifies authentication middleware
